+++ gakumas-tools/simulator/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to the contest simulator will be documented in this file.
 Dates are based on Eastern Time Zone.
 
+## 2025-05-23
+
+### Fixed
+
+- Held cards over limit (2) getting permanently removed instead of discarded
+
 ## 2025-05-19
 
 ### Fixed